% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_precision.R
\name{calculate_precision}
\alias{calculate_precision}
\title{\enumerate{
\item Calculate Model Precision
}}
\usage{
calculate_precision(actual, predicted)
}
\arguments{
\item{actual}{A numeric vector of actual outcomes (1s and 0s).}

\item{predicted}{A numeric vector of predicted outcomes (1s and 0s).}
}
\value{
A single numeric value representing the precision. If the denominator
is 0 (no positive predictions), the function should return NA.
}
\description{
Precision is defined as: True Positives / (True Positives + False Positives)
}
