% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_outliers.R
\name{cap_outliers}
\alias{cap_outliers}
\title{\enumerate{
\item Cap outliers using quantiles
}}
\usage{
cap_outliers(x, lower_quantile = 0.05, upper_quantile = 0.95)
}
\arguments{
\item{x}{A numeric vector.}

\item{lower_quantile}{A numeric value between 0 and 1. Default is 0.05.}

\item{upper_quantile}{A numeric value between 0 and 1. Default is 0.95.}
}
\value{
A numeric vector where values below the lower quantile are replaced
with the lower quantile value, and values above the upper quantile are
replaced with the upper quantile value.
}
\description{
\enumerate{
\item Cap outliers using quantiles
}
}
