% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_binary_target.R
\name{encode_binary_target}
\alias{encode_binary_target}
\title{\enumerate{
\item Encode a binary character/factor target variable into 0s and 1s
}}
\usage{
encode_binary_target(x, positive_class = "Yes")
}
\arguments{
\item{x}{A character or factor vector containing exactly two unique valid classes.}

\item{positive_class}{The string label that should be encoded as 1.}
}
\value{
A numeric vector of 1s and 0s.
}
\description{
\enumerate{
\item Encode a binary character/factor target variable into 0s and 1s
}
}
