% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_numeric_column.R
\name{impute_numeric_column}
\alias{impute_numeric_column}
\title{\enumerate{
\item Impute missing values in a numeric column
}}
\usage{
impute_numeric_column(df, column_name, method = "median")
}
\arguments{
\item{df}{A data frame.}

\item{column_name}{The name of the column to impute (as a string).}

\item{method}{A string, either "mean" or "median". Default is "median".}
}
\value{
A data frame with the specified column's missing values imputed.
}
\description{
\enumerate{
\item Impute missing values in a numeric column
}
}
